refactored special commands + api
